reprise paris

+++ db/migrate/20220722145723_add_montant_to_paris.rb
@@ -1,5 +1,5 @@
 class AddMontantToParis < ActiveRecord::Migration[7.0]
   def change
-    add_column :paris, :montant, :integer
+    add_column :paris, :montant, :decimal
   end
 end
