+++ db/migrate/20220722161857_add_solde_to_paris.rb
@@ -1,5 +1,5 @@
 class AddSoldeToParis < ActiveRecord::Migration[7.0]
   def change
-    add_column :paris, :solde, :integer
+    add_column :paris, :solde, :decimal
   end
 end
